Will only remove test location providers if the app has the ACCESS_MOCK_LOCATION permission

+++ ruby-gem/test-server/instrumentation-backend/src/sh/calaba/instrumentationbackend/InstrumentationBackend.java
@@ -1,5 +1,7 @@
 package sh.calaba.instrumentationbackend;
 
+import android.Manifest;
+import android.content.pm.PackageManager;
 import sh.calaba.instrumentationbackend.actions.Actions;
 import sh.calaba.instrumentationbackend.actions.HttpServer;
 import android.app.Activity;
@@ -71,11 +73,14 @@ public class InstrumentationBackend extends ActivityInstrumentationTestCase2 {
         Log.e(TAG, message);
     }
 
-    private static void removeTestLocationProviders(Activity activity) {
-        final LocationManager locationService =
-            (LocationManager) activity.getSystemService(Context.LOCATION_SERVICE);
-        for (final String provider : locationService.getAllProviders()) {
-            locationService.removeTestProvider(provider);
+    private void removeTestLocationProviders(Activity activity) {
+        int hasPermission = solo.getCurrentActivity().checkCallingOrSelfPermission(Manifest.permission.ACCESS_MOCK_LOCATION);
+
+        if (hasPermission == PackageManager.PERMISSION_GRANTED) {
+            LocationManager locationService = (LocationManager) activity.getSystemService(Context.LOCATION_SERVICE);
+            for (final String provider : locationService.getAllProviders()) {
+                locationService.removeTestProvider(provider);
+            }
         }
     }
 }
